Load Justifications from a file or multiple files

These function load justifications from the YAML fragments in one (load_justifications) or multiple files (load_justifications_dir).

load_justifications(
  text = NULL,
  file = NULL,
  delimiterRegEx = "^---$",
  justificationContainer = c("justifier", "justification", "decision", "assertion",
    "source"),
  ignoreOddDelimiters = FALSE,
  encoding = "UTF-8",
  silent = TRUE
)

load_justifications_dir(
  path,
  recursive = TRUE,
  extension = "jmd",
  regex = NULL,
  justificationContainer = c("justifier", "justification", "decision", "assertion",
    "source"),
  delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE,
  encoding = "UTF-8",
  silent = TRUE
)

Arguments

text, file
As text or file, you can specify a file to read with encoding encoding, which will then be read using base::readLines(). If the argument is named text, whether it is the path to an existing file is checked first, and if it is, that file is read. If the argument is named file, and it does not point to an existing file, an error is produced (useful if calling from other functions). A text should be a character vector where every element is a line of the original source (like provided by base::readLines()); although if a character vector of one element and including at least one newline character (\\n) is provided as text, it is split at the newline characters using base::strsplit(). Basically, this behavior means that the first argument can be either a character vector or the path to a file; and if you're specifying a file and you want to be certain that an error is thrown if it doesn't exist, make sure to name it file.

delimiterRegEx
The regular expression used to locate YAML fragments

justificationContainer
The container of the justifications in the YAML fragments. Because only justifications are read that are stored in this container, the files can contain YAML fragments with other data, too, without interfering with the parsing of the justifications.

ignoreOddDelimiters
Whether to throw an error (FALSE) or delete the last delimiter (TRUE) if an odd number of delimiters is encountered.

encoding
The encoding to use when calling readLines(). Set to NULL to let readLines() guess.

silent
Whether to be silent (TRUE) or informative (FALSE).

path
The path containing the files to read.

recursive
Whether to also process subdirectories (TRUE) or not (FALSE).

extension
The extension of the files to read; files with other extensions will be ignored. Multiple extensions can be separated by a pipe (|).

regex
Instead of specifing an extension, it's also possible to specify a regular expression; only files matching this regular expression are read. If specified, regex takes precedece over extension,

Value

An object with the ggplot2::ggplot graph stored in output$graph and the overview in output$overview.

Details

load_justifications_dir simply identifies all files and then calls load_justifications for each of them. load_justifications loads the YAML fragments containing the justifications using yum::load_yaml_fragments() and then parses the justifications into a visual representation as a ggplot2::ggplot graph and Markdown documents with overviews.

Examples

exampleMinutes <- 'This is an example of minutes that include
a source, an assertion, and a justification. For example, in
the meeting, we can discuss the assertion that sleep deprivation
affects decision making. We could quickly enter this assertion in
a machine-readable way in this manner:

---
assertion:
  -
    id: assertion_SD_decision
    label: Sleep deprivation affects the decision making proces.
    source:
      id: source_Harrison
---

Because it is important to refer to sources, we cite a source as well.
We have maybe specified that source elsewhere, for example in the
minutes of our last meeting. That specification may have looked
like this:

---
source:
  -
    id: source_Harrison
    label: "Harrison & Horne (2000) The impact of sleep deprivation on decision making: A review."
    xdoi: "doi:10.1037/1076-898x.6.3.236"
    type: "Journal article"
---

We can now refer to these two specifications later on, for
example to justify decisions we take.
';

justifier::load_justifications(text=exampleMinutes);

#> $raw
#> $assertion
#> $assertion$id
#> [1] "assertion_SD_decision"
#> 
#> $assertion$label
#> [1] "Sleep deprivation affects the decision making proces."
#> 
#> $assertion$source
#> $assertion$source$id
#> [1] "source_Harrison"
#> 
#> 
#> 
#> $source
#> $source$id
#> [1] "source_Harrison"
#> 
#> $source$label
#> [1] "Harrison & Horne (2000) The impact of sleep deprivation on decision making: A review."
#> 
#> $source$xdoi
#> [1] "doi:10.1037/1076-898x.6.3.236"
#> 
#> $source$type
#> [1] "Journal article"
#> 
#> 
#> attr(,"class")
#> [1] "simplifiedYum" "list"         
#> 
#> $structured
#> $structured$sources
#> $structured$sources$source_Harrison
#> $structured$sources$source_Harrison$id
#> [1] "source_Harrison"
#> 
#> $structured$sources$source_Harrison$label
#> [1] "Harrison & Horne (2000) The impact of sleep deprivation on decision making: A review."
#> 
#> $structured$sources$source_Harrison$xdoi
#> [1] "doi:10.1037/1076-898x.6.3.236"
#> 
#> $structured$sources$source_Harrison$type
#> [1] "Journal article"
#> 
#> $structured$sources$source_Harrison$fromFile
#> [1] "none"
#> 
#> $structured$sources$source_Harrison$fromDir
#> [1] "."
#> 
#> 
#> 
#> $structured$assertions
#> $structured$assertions$assertion_SD_decision
#> $id
#> [1] "assertion_SD_decision"
#> 
#> $label
#> [1] "Sleep deprivation affects the decision making proces."
#> 
#> $source
#> [1] "source_Harrison"
#> attr(,"class")
#> [1] "justifierRef"
#> 
#> $fromFile
#> [1] "none"
#> 
#> $fromDir
#> [1] "."
#> 
#> attr(,"class")
#> [1] "justifierSpec" "assertion"    
#> 
#> 
#> $structured$justifications
#> named list()
#> 
#> $structured$decisions
#> named list()
#> 
#> $structured$justifier
#> named list()
#> 
#> 
#> $logging
#> $logging$n_firstSweep
#>        sources     assertions justifications      decisions      justifier 
#>              1              1              0              0              0 
#> 
#> 
#> $frameworks
#> $frameworks$specs
#> list()
#> 
#> $frameworks$loaded
#> list()
#> 
#> 
#> $fwApplications
#> list()
#> 
#> $supplemented
#> $supplemented$sources
#> $supplemented$sources$source_Harrison
#> $supplemented$sources$source_Harrison$id
#> [1] "source_Harrison"
#> 
#> $supplemented$sources$source_Harrison$label
#> [1] "Harrison & Horne (2000) The impact of sleep deprivation on decision making: A review."
#> 
#> $supplemented$sources$source_Harrison$xdoi
#> [1] "doi:10.1037/1076-898x.6.3.236"
#> 
#> $supplemented$sources$source_Harrison$type
#> [1] "Journal article"
#> 
#> $supplemented$sources$source_Harrison$fromFile
#> [1] "none"
#> 
#> $supplemented$sources$source_Harrison$fromDir
#> [1] "."
#> 
#> 
#> 
#> $supplemented$assertions
#> $supplemented$assertions$assertion_SD_decision
#> $id
#> [1] "assertion_SD_decision"
#> 
#> $label
#> [1] "Sleep deprivation affects the decision making proces."
#> 
#> $source
#> $source$source_Harrison
#> $source$source_Harrison$id
#> [1] "source_Harrison"
#> 
#> $source$source_Harrison$label
#> [1] "Harrison & Horne (2000) The impact of sleep deprivation on decision making: A review."
#> 
#> $source$source_Harrison$xdoi
#> [1] "doi:10.1037/1076-898x.6.3.236"
#> 
#> $source$source_Harrison$type
#> [1] "Journal article"
#> 
#> $source$source_Harrison$fromFile
#> [1] "none"
#> 
#> $source$source_Harrison$fromDir
#> [1] "."
#> 
#> 
#> 
#> $fromFile
#> [1] "none"
#> 
#> $fromDir
#> [1] "."
#> 
#> attr(,"class")
#> [1] "justifierSpec" "assertion"    
#> 
#> 
#> $supplemented$justifications
#> named list()
#> 
#> $supplemented$decisions
#> named list()
#> 
#> $supplemented$justifier
#> named list()
#> 
#> 
#> $decisionTrees
#> named list()
#> attr(,"class")
#> [1] "justifierTree" "list"         
#> 
#> $decisionGraphs
#> named list()
#> 
#> attr(,"class")
#> [1] "justifications"

### To load a directory with justifications

examplePath <-
  file.path(system.file(package="justifier"),
            'extdata');
justifier::load_justifications_dir(path=examplePath);

#> Error in create_justifierGraph(x[[dTreeName]]): You must pass a justifier tree!